Cambios en el orden de los archivos, se agrego una linea a Customize, se agrego el export default a las views ruteadas para poder randerizar
