+++ Client/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
+import App from './components/App/App'
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux' 
 import store from '../redux/store.jsx'
