+++ Client/src/views/Customize/Customize.jsx
@@ -67,3 +67,5 @@ function CameraRig( { children } ){
 
     return <group ref={ group } >{ children }</group>
 }
+
+export default Customize
