+++ Client/src/views/Home.jsx
@@ -1,0 +1,10 @@
+import { NavLink } from "react-router-dom"
+
+const Home = () => {
+    return(
+        <>
+            <h1>Home</h1>
+        </>
+    )
+}
+export default Home;
