+++ Client/src/views/Login.jsx
@@ -1,0 +1,10 @@
+import { NavLink } from "react-router-dom"
+
+const Login = () => {
+    return(
+        <>
+            <h1>Custom Craft</h1>
+        </>
+    )
+}
+export default Login;
